Netflix _dashboard
The dashboard shows a comparison of the top performers on Netflix. It is broken down by:

Total movies & TV shows by country: A map of the world showing the count of Netflix movies and TV shows by country.

♦ #Top 10 actors by their hidden gem score: A bar chart showing the top 10 actors with the highest hidden gem score, which is a measure of how much people enjoy their performances.

♦ #Series and movies with their Awards: A pie chart that shows the breakdown of movies and TV shows on Netflix.

♦ #Top 10 actors with their IMDb score: A bar chart showing the top 10 actors with the highest IMDb score.

♦ #Writers with their Metacritic Score: A line graph showing the Metacritic scores of the top writers on Netflix.

This information can be used to identify the most popular performers on Netflix, as well as to see which countries have the most movies and TV shows available on the platform. It can also be used to see which writers are most highly regarded by critics.

Here are some additional points about the dashboard:

The dashboard is very visual and easy to understand.

The use of color is effective and helps to highlight important information.

The dashboard is interactive and allows users to explore the data in more detail.

Overall, this is a well-designed dashboard that provides a comprehensive overview of Netflix's top performers.